Added HTTP status check in aura_auth.go
Added defer resp.Body.Close() in aura_auth.go and aura_client.go to avoid memory leaks

Adjusted logic to detect auth token has expires.  Auth endpoint returns the duration that a token is valid for rather than the date / time it expires.

Adjusted aura_response.go TokenResponse to hold TokenExpires that is a calculated time/date, in Unix format , when the token will expiry.  The logic checks to see if the current time/date is within 60 seconds of that; if it is, a new token is fetched.